Applied code review
