+++ fbw-a32nx/src/wasm/systems/a320_systems/src/lib.rs
@@ -29,10 +29,10 @@ use systems::simulation::InitContext;
 use uom::si::{f64::Length, length::nautical_mile};
 
 use systems::{
+    air_starter_unit::AirStarterUnit,
     apu::{
-        AirStarterUnit, Aps3200ApuGenerator, Aps3200StartMotor, AuxiliaryPowerUnit,
-        AuxiliaryPowerUnitFactory, AuxiliaryPowerUnitFireOverheadPanel,
-        AuxiliaryPowerUnitOverheadPanel,
+        Aps3200ApuGenerator, Aps3200StartMotor, AuxiliaryPowerUnit, AuxiliaryPowerUnitFactory,
+        AuxiliaryPowerUnitFireOverheadPanel, AuxiliaryPowerUnitOverheadPanel,
     },
     electrical::{Electricity, ElectricitySource, ExternalPowerSource},
     engine::{leap_engine::LeapEngine, reverser_thrust::ReverserForce, EngineFireOverheadPanel},
